# migrate `gather` to ATen (CUDA)