Fixes [#24567](https://github.com/pytorch/pytorch/issues/24567).