@nikitaved 2020-05-01
Benchmarking:
### Case 1: many-to-one/one-to-many
```python
import torch
from IPython import get_ipython

ipython = get_ipython()

torch.set_num_threads(1)

device = 'cuda'

nfeat = 10000
nrep = 5
a=torch.arange(nfeat, device=device).repeat_interleave(nrep)
batch=3 #batch can vary 1-200

for batch in [100, 500, 1000, 5000, 10000]:
    print("Batch: ", batch)
    c=torch.randint(3, (batch, nfeat * nrep), device=device).float()
    ipython.magic("timeit c.gather(1,a.unsqueeze(0).expand(batch,a.size(0))); torch.cuda.synchronize()")

    enum_values = [
        list(range(1, 201)),
        list(range(1000, 1020)),
        list(range(2000, 2010)),
        list(range(3000, 3206)),
    ]
    indices = torch.tensor([i for i, values in enumerate(enum_values) for _j in range(len(values))], device=device)
    c = torch.randint(3, (batch, 4), device=device).float()
    idx = indices.unsqueeze(0).expand(c.size(0), indices.size(0))
    src = c.repeat(1, idx.shape[-1] // c.shape[-1])
    ipython.magic("timeit src.gather(1,idx); torch.cuda.synchronize()")
    print()

```
#### Master
```
Batch:  100                                                                                                                                       
83.2 µs ± 218 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)                                                                          
14 µs ± 74.7 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)                                                                          
                                                                                                                                                  
Batch:  500                                                                                                                                       
347 µs ± 1.11 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)                                                                           
15.7 µs ± 177 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)                                                                         
                                                                                                                                                  
Batch:  1000                                                                                                                                      
683 µs ± 1.54 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)                                                                           
16.2 µs ± 82 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)                                                                          
                                                                                                                                                  
Batch:  5000                                                                                                                                      
3.37 ms ± 7.86 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)                                                                           
37.3 µs ± 76.9 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

Batch:  10000
6.79 ms ± 2.61 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
62 µs ± 35.1 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

```
#### This PR
```
Batch:  100
75 µs ± 254 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
14.7 µs ± 66.6 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

Batch:  500
301 µs ± 129 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
15.5 µs ± 27.3 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

Batch:  1000
585 µs ± 647 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
16.6 µs ± 272 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

Batch:  5000
2.86 ms ± 9.25 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
33 µs ± 28.5 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

Batch:  10000
5.73 ms ± 5.44 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
53.3 µs ± 50.7 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)

```

### Case 1: sparse indexing
```python
import torch
from IPython import get_ipython

ipython = get_ipython()

torch.set_num_threads(1)

nrows = 100000
ncols = 100000
dims = [nrows, ncols]

src = torch.randn(dims)
idx1 = torch.randint(dims[0], (1, dims[1])).long()
idx2 = torch.randint(dims[1], (dims[0], 1)).long()

ipython.magic("timeit src.gather(0, idx1)")
ipython.magic("timeit src.gather(1, idx2)")

dims = dims[::-1]

src = torch.randn(dims)
idx1 = torch.randint(dims[0], (1, dims[1])).long()
idx2 = torch.randint(dims[1], (dims[0], 1)).long()

ipython.magic("timeit src.gather(0, idx1)")
ipython.magic("timeit src.gather(1, idx2)")
```
#### Master
```
4.82 ms ± 2.31 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
6.4 ms ± 42.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
4.65 ms ± 11 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
5.97 ms ± 13.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

```
#### This PR
```
4.93 ms ± 20 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
6.4 ms ± 14.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
4.61 ms ± 24.8 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
5.9 ms ± 19.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

```
#### Case 3: dense indexing
```python
import torch
import numpy
from IPython import get_ipython

numpy.random.seed(13)
torch.manual_seed(13)

ipython = get_ipython()

Ms=1024 * 8
Ns=1024 * 4
dim = 0
top_power = 4

for pM in range(top_power):
    M = Ms * (2 ** pM)
    for pN in range(top_power):
        N = Ns * (2 ** pN)
        src = torch.rand(M, N, device=torch.device('cuda'))
        index = torch.tensor(numpy.random.randint(0, min(M, N), (M, N)), device=torch.device('cuda') )
        print(f"Problem size (MxN): {M}x{N}")
        ipython.magic("timeit src.gather(0, index); torch.cuda.synchronize()")
        ipython.magic("timeit src.gather(1, index); torch.cuda.synchronize()")

```
#### Master
```
Problem size (MxN): 8192x4096                                                                                                                     
8.96 ms ± 2.54 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)                                                                           
946 µs ± 620 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)                                                                            
Problem size (MxN): 8192x8192                                                                                                                     
19.9 ms ± 1.11 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)                                                                           
1.88 ms ± 135 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)                                                                           
Problem size (MxN): 8192x16384                                                                                                                    
40.3 ms ± 2.49 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                            
3.49 ms ± 11 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)                                                                             
Problem size (MxN): 8192x32768                                                                                                                    
81.1 ms ± 2.66 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                            
6.71 ms ± 12.1 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)                                                                           
Problem size (MxN): 16384x4096                                                                                                                    
18 ms ± 1.16 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)                                                                             
1.87 ms ± 310 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)                                                                           
Problem size (MxN): 16384x8192                                                                                                                    
39.7 ms ± 3.03 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                            
3.78 ms ± 2.56 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)                                                                           
Problem size (MxN): 16384x16384                                                                                                                   
81.3 ms ± 2.64 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                            
7.63 ms ± 3.28 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)                                                                           
Problem size (MxN): 16384x32768                                                                                                                   
163 ms ± 5.29 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                             
14 ms ± 6.49 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)                                                                             
Problem size (MxN): 32768x4096
35.8 ms ± 3.04 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
3.74 ms ± 618 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 32768x8192
79.4 ms ± 4.22 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
7.51 ms ± 1.17 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 32768x16384
163 ms ± 4.23 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
15.3 ms ± 3.94 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 32768x32768
327 ms ± 16.2 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
31.8 ms ± 121 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 65536x4096
71.6 ms ± 4.74 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
7.47 ms ± 5.59 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 65536x8192
159 ms ± 3.11 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
15.1 ms ± 1.96 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 65536x16384
325 ms ± 19.8 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
30.5 ms ± 73.2 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 65536x32768
655 ms ± 12.6 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
72.4 ms ± 48.3 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

```
#### This PR
```
Problem size (MxN): 8192x4096                                                                                                           [328/1812]
9.34 ms ± 1.36 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
943 µs ± 960 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
Problem size (MxN): 8192x8192
20.4 ms ± 1.15 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
1.87 ms ± 704 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
Problem size (MxN): 8192x16384
41.5 ms ± 1.16 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
3.29 ms ± 511 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 8192x32768
83.5 ms ± 4.26 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
6.13 ms ± 684 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 16384x4096
18.7 ms ± 642 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
1.86 ms ± 821 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
Problem size (MxN): 16384x8192
40.9 ms ± 2.94 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
3.73 ms ± 845 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 16384x16384
83.5 ms ± 2.32 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
7.51 ms ± 502 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 16384x32768
168 ms ± 4.71 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
13.4 ms ± 1.95 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 32768x4096
37.4 ms ± 2.98 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
3.71 ms ± 2.08 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 32768x8192
81.7 ms ± 2.2 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
7.43 ms ± 270 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 32768x16384
167 ms ± 4.09 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
15 ms ± 1.09 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 32768x32768
336 ms ± 24.4 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
30.2 ms ± 22.1 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 65536x4096
74.8 ms ± 2.12 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
7.4 ms ± 1.21 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 65536x8192
163 ms ± 2.01 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
14.9 ms ± 613 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
Problem size (MxN): 65536x16384
334 ms ± 18.6 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
30 ms ± 16.1 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
Problem size (MxN): 65536x32768
672 ms ± 52.9 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
60.3 ms ± 18.6 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

```